+++ CHANGELOG.md
@@ -989,3 +989,4 @@
 - [2026-05-09 21:27] feat(navigation): implement shimmer loading skeleton in navigation
 - [2026-05-10 09:15] feat(navigation): implement shimmer loading skeleton in navigation
 - [2026-05-10 09:18] feat(i18n): add bi-directional event bus handler for i18n
+- [2026-05-10 09:53] feat(security): implement responsive security view component